requirements.txt
numpy<2
Pillow
opencv-python
PyYAML
ultralytics
